window: Fix signedness warning

+++ src/core/window.c
@@ -10002,7 +10002,7 @@ meta_window_handle_mouse_grab_op_event (MetaWindow         *window,
     {
     case CLUTTER_BUTTON_RELEASE:
       if (event->button.button == 1 ||
-          event->button.button == meta_prefs_get_mouse_button_resize ())
+          event->button.button == (unsigned int) meta_prefs_get_mouse_button_resize ())
         {
           meta_display_check_threshold_reached (window->display,
                                                 event->button.x,
